Simple payments block: re-organise utils
